(res): Resolved Auth Bugs and added validation , error messages handling

+++ backend/controllers/auth.controller.js
@@ -7,25 +7,28 @@ import { generateToken, verifyToken } from "../utils/jwt.js";
 import { sendOTPMail } from "../utils/nodemailer.js";
 import { generateOtp } from "../utils/otp.js";
 import { printKeyValuePair } from "../utils/redis-debug.js";
-import axios from 'axios'
+import axios from "axios";
 import { userValidator } from "../validators/userValidator.js";
+import { response } from "express";
+import { chat } from "googleapis/build/src/apis/chat/index.js";
 
 const OTP_INVALID_TIME = process.env.OTP_INVALID_TIME || 120;
 const OTP_EXPIRY_TIME = process.env.OTP_EXPIRY_TIME || 300;
 const OTP_VERIFY_LIMIT = process.env.OTP_VERIFY_LIMIT || 3;
 
+// Handles user registeration , If user unverified it deletes and recreates
 export const registerUser = async (req, res, next) => {
     try {
-        const {value, error} = userValidator(req.body)
-        if(error){
-            res.status(400)
-            throw error
+        const { value, error } = userValidator(req.body);
+        if (error) {
+            res.status(400);
+            throw error;
         }
         const { email } = req.body;
 
         const user = await User.findOne({ email: email });
         console.log(user, email);
-        
+
         if (user && user.is_verified) {
             console.log(1);
             res.status(400);
@@ -33,11 +36,12 @@ export const registerUser = async (req, res, next) => {
         } else if (user) {
             await User.findOneAndDelete({ email: email });
         }
-        
+
+        await User.create({ ...req.body, is_verified: false });
 
         const otp = generateOtp();
         //senting otp and saving in redis cache
-        const data = await Promise.all([
+        await Promise.all([
             client.setEx(
                 `OTP:${email}`,
                 OTP_EXPIRY_TIME,
@@ -52,14 +56,13 @@ export const registerUser = async (req, res, next) => {
             ),
             sendOTPMail(email, otp),
         ]);
-        
-        await User.create({ ...req.body, is_verified: false });
-        console.log(data);
-        console.log(await User.findOne({ email }));
+
+        console.log(OTP_EXPIRY_TIME, OTP_INVALID_TIME);
 
         return res.json({
             message: "OTP sent successfully",
             status: "success",
+            data: { otp_timer: Date.now() + Number(OTP_INVALID_TIME * 1000) },
         });
     } catch (error) {
         console.log(error);
@@ -71,11 +74,13 @@ export const loginUser = async (req, res, next) => {
     try {
         const { email, password } = req.body;
         const user = await User.findOne({ email });
-        if (!user) {
+        if (!user || !user.is_verified) {
             res.status(404);
             throw new Error("User not found");
         }
+
         console.log(email, password);
+
         const isPasswordCorrect = await user.comparePassword(password);
         if (!isPasswordCorrect) {
             res.status(400);
@@ -129,7 +134,7 @@ export const verifyOTP = async (req, res, next) => {
             console.log("expirye");
 
             res.status(404);
-            return next("OTP expired or doesnt exists");
+            throw new Error("OTP expired or doesnt exists");
         }
 
         //*getting otp from redis cache
@@ -138,29 +143,40 @@ export const verifyOTP = async (req, res, next) => {
         console.log(otp, otp_count, otp == otp_code, otp_code);
         console.log(cache_data);
 
+        if (type !== otp_type) {
+            await client.del(`OTP${email}`);
+            res.status(400);
+            throw new Error("Invalid OTP verification attempt");
+        }
+
         //* checking if otp timeout
         if (Date.now() > otp_invalid_time) {
-            await client.del(`OTP:${email}`);
+            // await client.del(`OTP:${email}`);
             res.status(400);
-            return next(new Error("OTP Timeout"));
+            throw new Error("OTP Timeout");
         }
 
         //* Checking if limit exceeded
         if (otp_count >= OTP_VERIFY_LIMIT) {
             await client.del(`OTP:${email}`);
             res.status(400);
-            return next(new Error("OTP count limit reached, Resend OTP"));
+            throw new Error("OTP count limit reached, Resend OTP");
         }
         //res on invalid otp
         if (otp != otp_code) {
             console.log(123);
             await client.set(
                 `OTP:${email}`,
-                JSON.stringify({ otp, otp_count: otp_count + 1, otp_type }),
+                JSON.stringify({
+                    otp,
+                    otp_count: otp_count + 1,
+                    otp_type,
+                    otp_invalid_time,
+                }),
                 "KEEPTTL"
             );
             res.status(400);
-            return next(new Error("Invalid OTP"));
+            throw new Error("Invalid OTP");
         }
         await client.del(`OTP:${email}`);
         console.log(cache_data);
@@ -169,7 +185,7 @@ export const verifyOTP = async (req, res, next) => {
             const oldUser = await User.findOne({ email });
             if (!oldUser) {
                 res.status(404);
-                return next(new Error("User not found"));
+                throw new Error("User not found");
             }
             const newUser = await User.findOneAndUpdate(
                 { email: email },
@@ -200,12 +216,20 @@ export const verifyOTP = async (req, res, next) => {
 export const resendOTP = async (req, res, next) => {
     printKeyValuePair();
     try {
-        const { email, otp, otp_type } = req.body;
+        const { email } = req.body;
         const cache = JSON.parse(await client.get(`OTP:${email}`));
-        if (cache) {
-            await client.del(`OTP:${email}`);
+        if (!cache) {
+            throw new Error("No OTP found");
         }
 
+        if (Date.now() < cache.otp_invalid_time) {
+            res.status(400);
+            throw new Error("No Resend until invalid time");
+        }
+
+        await client.del(`OTP:${email}`);
+
+        console.log(cache);
         const otp_code = generateOtp();
         console.log(cache, 123);
         //senting otp and saving in redis cache
@@ -226,20 +250,25 @@ export const resendOTP = async (req, res, next) => {
         ]);
 
         res.status(200);
-        res.json({ message: "OTP resend successful", status: "success" });
+        res.json({
+            message: "OTP resend successful",
+            status: "success",
+            data: { otp_timer: Date.now() + Number(OTP_INVALID_TIME * 1000) },
+        });
     } catch (error) {
         next(error);
     }
 };
 
 export const forgetPassword = async (req, res, next) => {
+    printKeyValuePair()
     try {
         const { email } = req.body;
         console.log(1);
         const user = await User.findOne({ email, is_verified: true });
         if (!user) {
             res.status(400);
-            return next(new Error("No User found with this email Id"));
+            throw new Error("No User found with this email Id");
         }
         const otp_code = generateOtp();
         //senting otp and saving in redis cache
@@ -251,6 +280,9 @@ export const forgetPassword = async (req, res, next) => {
                     otp: otp_code,
                     otp_count: 0,
                     otp_type: OTP_TYPES.FORGET_PASSWORD,
+                    otp_invalid_time:
+                        new Date(Date.now()).getTime() +
+                        OTP_INVALID_TIME * 1000,
                 })
             ),
             sendOTPMail(email, otp_code),
@@ -259,7 +291,7 @@ export const forgetPassword = async (req, res, next) => {
         res.status(200).json({
             message: "OTP send successfully",
             status: "success",
-            data: { otp_timer: Date.now() + Number(OTP_EXPIRY_TIME * 1000) },
+            data: { otp_timer: Date.now() + Number( OTP_INVALID_TIME * 1000) },
         });
     } catch (error) {
         next(error);
@@ -269,23 +301,29 @@ export const forgetPassword = async (req, res, next) => {
 export const resetPassword = async (req, res, next) => {
     await printKeyValuePair();
     try {
-        const { email, password } = req.body;
+        const { email, password , confirm_password } = req.body;
         const cache = await client.get(`otp_verified:${email}`);
         console.log(cache, 1234, req.body);
+        if(password !== confirm_password){
+            res.status(400)
+            throw new Error("Password not same")
+        }
+        
         if (cache === "true") {
+            
             const user = await User.findOne({ email: email });
-            if(!user){
-                throw new Error("User not found")
+            if (!user) {
+                throw new Error("User not found");
             }
-            user.password = password
-            await user.save()
-            console.log(user)
+            user.password = password;
+            await user.save();
+            console.log(user);
             await client.del(`otp_verified:${email}`);
         } else {
             res.status(400);
-            return next(new Error("OTP not verified"));
+            throw new Error("Unauthorized");
         }
-        return res.json({ status: "success" });
+        return res.json({ status: "success" , message : "Password Reset Successful"});
     } catch (error) {
         next(error);
     }
@@ -295,12 +333,12 @@ export const getUserDetails = async (req, res, next) => {
     try {
         const token = req.cookies.jwt;
         if (!token) {
-            res.status(403);
+            res.status(401);
             return next("Access Denied");
         }
         const data = verifyToken(req.cookies.jwt);
         const user = await User.findOne({ email: data.email });
-        console.log(user)
+        console.log(user);
         if (!user || user.active === false) {
             res.status(403);
             return next("Access Denied");
@@ -316,32 +354,44 @@ export const getUserDetails = async (req, res, next) => {
             status: "success",
         });
     } catch (error) {
-        res.status(403);
+        res.status(401);
         next(error);
     }
 };
 
-export const googleAuth = async(req, res, next) => {
-    const {code} = req.query
-    try{
-        const googleRes = await oauth2Client.getToken(code)
-        oauth2Client.setCredentials(googleRes.tokens)
-        const userRes = await axios.get(`https://www.googleapis.com/oauth2/v1/userinfo?alt=json&access_token=${googleRes.tokens.access_token}`)
-        console.log(userRes)
-        const {email , name } = userRes.data
-        
-        const [first_name , last_name] = name.split(' ')
-        let user = await User.findOne({email})
-        
-        if(!user){
-            user = await User.create({first_name , last_name , email})
+export const googleAuth = async (req, res, next) => {
+    const { code } = req.query;
+    try {
+        const googleRes = await oauth2Client.getToken(code);
+        oauth2Client.setCredentials(googleRes.tokens);
+        const userRes = await axios.get(
+            `https://www.googleapis.com/oauth2/v1/userinfo?alt=json&access_token=${googleRes.tokens.access_token}`
+        );
+        console.log(userRes);
+        const { email, name } = userRes.data;
+
+        const [first_name, last_name] = name.split(" ");
+        let user = await User.findOne({ email });
+
+        if (!user) {
+            user = await User.create({
+                first_name,
+                last_name,
+                email,
+                is_verified: true,
+            });
         }
-        const {_id} = user
-        const token = generateToken(email)
-        res.status(200).cookie('jwt', token).json({status : "success", message : "Google Login Success", data : user})
-        
-    }catch(error){
-        console.log(error)
-        next(error)
+        const { _id } = user;
+        const token = generateToken(email);
+        res.status(200)
+            .cookie("jwt", token, { maxAge: 24 * 60 * 60 * 1000 })
+            .json({
+                status: "success",
+                message: "Google Login Success",
+                data: user,
+            });
+    } catch (error) {
+        console.log(error);
+        next(error);
     }
-}
+};
